[Refactor] New package named pref_utils
