+++ app/src/main/java/com/twoday/todaytrip/pref_utils/PageNoPrefUtil.kt
@@ -1,4 +1,4 @@
-package com.twoday.todaytrip.utils
+package com.twoday.todaytrip.pref_utils
 
 import android.content.Context
 import android.content.SharedPreferences
