+++ app/src/main/java/com/twoday/todaytrip/ui/place_list/adapter/OnTourItemAddClickListener.kt
@@ -1,7 +1,7 @@
 package com.twoday.todaytrip.ui.place_list.adapter
 
 import com.twoday.todaytrip.tourData.TourItem
-import com.twoday.todaytrip.utils.ContentIdPrefUtil
+import com.twoday.todaytrip.pref_utils.ContentIdPrefUtil
 
 object OnTourItemAddClickListener {
     fun onTourItemAddClick(tourItem:TourItem){
