+++ app/src/main/java/com/twoday/todaytrip/ui/route/ItemTouchSimpleCallback.kt
@@ -3,9 +3,8 @@ package com.twoday.todaytrip.ui.route
 import android.util.Log
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
-import com.twoday.todaytrip.utils.ContentIdPrefUtil
+import com.twoday.todaytrip.pref_utils.ContentIdPrefUtil
 import java.lang.Integer.max
-import java.util.Collections
 
 interface OnMoveEndListener{
     fun onMoveEnd()
