+++ app/src/main/java/com/twoday/todaytrip/viewModel/RandomResultViewModel.kt
@@ -6,15 +6,14 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.twoday.todaytrip.R
 import com.twoday.todaytrip.tourApi.TourNetworkInterfaceUtils
-import com.twoday.todaytrip.utils.ContentIdPrefUtil
+import com.twoday.todaytrip.pref_utils.ContentIdPrefUtil
 import com.twoday.todaytrip.utils.DestinationData
-import com.twoday.todaytrip.utils.DestinationPrefUtil
-import com.twoday.todaytrip.utils.PageNoPrefUtil
-import com.twoday.todaytrip.utils.RecommendPrefUtil
-import com.twoday.todaytrip.utils.TourItemPrefUtil
+import com.twoday.todaytrip.pref_utils.DestinationPrefUtil
+import com.twoday.todaytrip.pref_utils.PageNoPrefUtil
+import com.twoday.todaytrip.pref_utils.RecommendPrefUtil
+import com.twoday.todaytrip.pref_utils.TourItemPrefUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
