+++ app/src/main/java/com/twoday/todaytrip/viewModel/RandomThemeViewModel.kt
@@ -3,7 +3,7 @@ package com.twoday.todaytrip.viewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.twoday.todaytrip.utils.DestinationPrefUtil
+import com.twoday.todaytrip.pref_utils.DestinationPrefUtil
 
 class RandomThemeViewModel : ViewModel() {
